+++ lib/cext/websocket_driver_patches.rb
@@ -1,0 +1,15 @@
+class WebsocketDriverPatches < CommonPatches
+
+  PATCHES = {
+    'websocket_mask.c' => {
+      gem: 'websocket-driver',
+      patches: [
+        {
+          match: /(VALUE .*?)\s+= Qnil;/,
+          replacement: '\1;'
+        }
+      ]
+    },
+  }
+  
+end
